Use assert() from libc instead of rolling our own
Using what is provided by the C library of the system should always be
preferred, as long as it's know to work fine. There's no problem with
assert(), so just use it.

Note that the rockspec will add -DNDEBUG by default, which is fine for
release builds. Developers might want to disable the define there when
debugging.